UVS-UNet/requirements.txt
matplotlib
numpy
pillow
scipy
torch==1.10.0
tqdm